Converted .Rd format to markdown, remove examples
Roxygen Markdown was alwas enabled in this package, but never used. This actually led to errors where the documentation could not be built, in contrast to the mogsa package, which this is based on, where markdown support was not enabled in roxygen2. I have converted the .Rd markup to markdown using the roxygen2md::roxygen2md() function.

Also removed examples (for now), as they are mostly outdated.